feat: moving admin endpoints to label controler. Auto reloading labels on frontend after updating them
