+++ src_back/controllers/AdminController.ts
@@ -4,11 +4,18 @@ import * as path from "path";
 import Config from '../utils/Config.js';
 import AbstractController from "./AbstractController.js";
 import Logger from "../utils/Logger.js";
+import Utils from "../utils/Utils.js";
+import SSEController, { SSECode } from "./SSEController.js";
+import {spawn} from "child_process";
 
 /**
 * Created : 14/12/2022 
 */
 export default class AdminController extends AbstractController {
+
+	public static announcements_cache:string = "[]";
+	private savingLabels:boolean = false;
+	private savingPromise:Promise<void>|null = null;
 	
 	constructor(public server:FastifyInstance) {
 		super();
@@ -33,6 +40,23 @@ export default class AdminController extends AbstractController {
 		this.server.delete('/api/admin/premium', async (request, response) => await this.deletePremium(request, response));
 		this.server.delete('/api/admin/beta/user', async (request, response) => await this.delUser(request, response));
 		this.server.delete('/api/admin/beta/user/all', async (request, response) => await this.removeAllUsers(request, response));
+
+		//Updates labels
+		this.server.post('/api/admin/labels', async (request:FastifyRequest, response:FastifyReply) => this.postLabels(request, response) );
+		
+		//Forces labels reload from frontend
+		this.server.post('/api/admin/labels/reload', async (request:FastifyRequest, response:FastifyReply) => this.postLabelsReload(request, response) );
+
+		//Add an annoucement
+		this.server.post('/api/admin/announcements', async (request:FastifyRequest, response:FastifyReply) => this.postAnnouncement(request, response) );
+
+		//Add an annoucement
+		this.server.delete('/api/admin/announcements', async (request:FastifyRequest, response:FastifyReply) => this.deleteAnnouncement(request, response) );
+
+		//Initialize cache
+		if(fs.existsSync(Config.ANNOUNCEMENTS_PATH)) {
+			AdminController.announcements_cache = fs.readFileSync(Config.ANNOUNCEMENTS_PATH, "utf-8");
+		}
 	}
 	
 	
@@ -235,4 +259,200 @@ export default class AdminController extends AbstractController {
 		response.status(200);
 		response.send(JSON.stringify({success:true, uids}));
 	}
+
+	/**
+	 * Adds an announcement
+	 * @param request 
+	 * @param response 
+	 * @returns 
+	 */
+	private postAnnouncement(request:FastifyRequest, response:FastifyReply):void {
+		if(!super.adminGuard(request, response)) return;
+
+		const body:any = request.body;
+		const dateStart:number = body.dateStart || Date.now();
+		const dateEnd:number = body.dateEnd;
+		const important:boolean = body.important === true;
+		const title:{[key:string]:string} = body.title;
+		const text:{[key:string]:string} = body.text;
+		const versionMax:string = body.versionMax || "";
+		const donorsOnly:boolean = body.donorsOnly === true;
+		const premiumOnly:boolean = body.premiumOnly === true;
+		const patreonOnly:boolean = body.patreonOnly === true;
+		const heatOnly:boolean = body.heatOnly === true;
+
+		let list:AnnouncementData[] = [];
+		if(fs.existsSync(Config.ANNOUNCEMENTS_PATH)) {
+			list = JSON.parse(fs.readFileSync(Config.ANNOUNCEMENTS_PATH, "utf-8")) as AnnouncementData[];
+		}
+
+		const announce:AnnouncementData ={
+			id:Utils.getUUID(),
+			dateStart,
+			title,
+			text,
+			important,
+			donorsOnly,
+			premiumOnly,
+			patreonOnly,
+			heatOnly,
+		};
+		if(dateEnd) announce.dateEnd = dateEnd;
+		if(versionMax.length >= 2) announce.versionMax = versionMax;
+		list.push(announce);
+
+		AdminController.announcements_cache = JSON.stringify(list);
+		fs.writeFileSync(Config.ANNOUNCEMENTS_PATH, AdminController.announcements_cache, "utf-8");
+
+		response.header('Content-Type', 'application/json');
+		response.status(200);
+		response.send(JSON.stringify({success:true, data:list}));
+	}
+
+	/**
+	 * Deletes an announcement
+	 */
+	private deleteAnnouncement(request:FastifyRequest, response:FastifyReply):void {
+		if(!super.adminGuard(request, response)) return;
+
+		const body:any = request.body;
+		const id:string = body.id;
+
+		let list:AnnouncementData[] = [];
+		if(fs.existsSync(Config.ANNOUNCEMENTS_PATH)) {
+			list = JSON.parse(fs.readFileSync(Config.ANNOUNCEMENTS_PATH, "utf-8")) as AnnouncementData[];
+		}
+
+		const index = list.findIndex(v=>v.id === id);
+		if(index > -1) {
+			list.splice(index, 1);
+		}
+
+		AdminController.announcements_cache = JSON.stringify(list);
+		fs.writeFileSync(Config.ANNOUNCEMENTS_PATH, AdminController.announcements_cache, "utf-8");
+
+		response.header('Content-Type', 'application/json');
+		response.status(200);
+		response.send(JSON.stringify({success:true, data:list}));
+	}
+
+	/**
+	 * Forces labels reload from frontend
+	 * @param request 
+	 * @param response 
+	 * @returns 
+	 */
+	public async postLabelsReload(request:FastifyRequest, response:FastifyReply):Promise<void> {
+		const body:any = request.body;
+		if(body.key != Config.credentials.csrf_key) {
+			response.header('Content-Type', 'application/json');
+			response.status(401);
+			response.send(JSON.stringify({success:false}));
+			return;
+		}
+		
+		Config.credentials.admin_ids.forEach(uid => {
+			SSEController.sendToUser(uid, SSECode.LABELS_UPDATE);
+		})
+		
+		response.header('Content-Type', 'application/json');
+		response.status(200);
+		response.send(JSON.stringify({success:true}));
+	}
+
+	/**
+	 * Updates a labels section
+	 * @param request 
+	 * @param response 
+	 * @returns 
+	 */
+	public async postLabels(request:FastifyRequest, response:FastifyReply):Promise<void> {
+		const user = await this.adminGuard(request, response);
+		if(!user) return;
+		
+		const body:any = request.body;
+		const lang:string = body.lang;
+		const json:any = {};
+		const section:string = body.section;
+		json[section] = body.labels;
+
+		if(this.savingLabels) {
+			Logger.warn("A label update is already in progress, wait for it complete before updateing section",section);
+			await this.savingPromise;
+			return await this.postLabels(request, response);
+		}
+
+		this.savingLabels = true;
+
+		let globalResolver!:()=>void;
+		this.savingPromise = new Promise((resolve) => {
+			globalResolver = resolve;
+		});
+		
+		Logger.info("Updating labels section \""+section+"\"");
+
+		try {
+			await new Promise<void>(resolve => {
+				fs.writeFileSync(path.join(Config.LABELS_ROOT, lang+"/"+section+".json"), JSON.stringify(json, null, "\t"), "utf-8");
+				const script = path.join(Config.LABELS_ROOT, "../src_labels/index.mjs");
+				// exec(`start /B node ${script}`, (error, stdout, stderr) => {
+				let errored = false;
+				const child = spawn('node', [script], {shell: false, windowsHide:true});
+				
+				child.stdout.on('data', data => {
+					console.log(data.toString());
+				});
+				
+				child.stderr.on('data', data => {
+					console.error(data.toString());
+					errored = true;
+				});
+	
+				child.on('error', (error) => {
+					console.error(error.message);
+					errored = true;
+				});
+				
+				child.on('close', (code) => {
+					if (errored) {
+						Logger.error("Failed compiling labels");
+						response.header('Content-Type', 'application/json');
+						response.status(500);
+						response.send(JSON.stringify({success:false, message:"failed executing script to compile labels", errorCode:"EXEC_SCRIPT_ERROR"}));
+						resolve();
+						return;
+					}
+					Logger.success("Labels compiled successfuly");
+					response.header('Content-Type', 'application/json');
+					response.status(200);
+					response.send(JSON.stringify({success:true}));
+					this.savingLabels = false;
+					resolve();
+					SSEController.sendToUser(user.user_id, SSECode.LABELS_UPDATE);
+				});
+			});
+
+		}catch(error) {
+			response.header('Content-Type', 'application/json');
+			response.status(500);
+			response.send(JSON.stringify({success:false, message:"an unknown error has occured", errorCode:"UNKNOWN_ERROR"}));
+		}
+		this.savingLabels = false;
+		globalResolver();
+	}
+
+}
+
+interface AnnouncementData{
+	id:string;
+	dateStart:number;
+	important:boolean;
+	donorsOnly:boolean;
+	premiumOnly:boolean;
+	patreonOnly:boolean;
+	heatOnly:boolean;
+	title:{[key:string]:string};
+	text:{[key:string]:string};
+	dateEnd?:number;
+	versionMax?:string;
 }
